use int stead of byte

+++ src/SHT3X.cpp
@@ -2,7 +2,7 @@
 
 namespace SHT3X
 {
-  byte SHT3X::read()
+  int SHT3X::read()
   {
     uint8_t data[6];
